Enable/disable message tray integration in schema.
Adjust the schema to provide an additional checkbox that allows the user
to disable or enable the message tray integration. Since the purpose of
this extension is primarily to provide integration with the gnome shell
message tray, I made the default on :-) I am proposing this change to
provide an option to users that instead use this extension mainly for
the pidgin buddy search provider.